+++ packages/bundle/src/createDirectLineAppServiceExtension.js
@@ -1,10 +1,12 @@
 import { DirectLineStreaming } from 'botframework-directlinejs';
 
-export default async function createDirectLineAppServiceExtension({ botAgent, conversationId, domain, token }) {
-  return new DirectLineStreaming({
-    botAgent,
-    conversationId,
-    domain,
-    token
-  });
+export default function createDirectLineAppServiceExtension({ botAgent, conversationId, domain, token }) {
+  return Promise.resolve(
+    new DirectLineStreaming({
+      botAgent,
+      conversationId,
+      domain,
+      token
+    })
+  );
 }