+++ packages/component/src/Middleware/ActivityStatus/createTimestampMiddleware.js
@@ -5,7 +5,7 @@ import Timestamp from './Timestamp';
 export default function createTimestampMiddleware() {
   return () => next => ({ activity, sameTimestampGroup, ...args }) => {
     if (!sameTimestampGroup) {
-      return <Timestamp activity={activity} aria-hidden={true} />;
+      return <Timestamp activity={activity} />;
     }
 
     return next({ activity, sameTimestampGroup, ...args });